Exceptions commit

+++ Digital-Library-Management-System/src/main/java/kadri/Digital/Library/Management/System/service/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package kadri.Digital.Library.Management.System.service;
 
 import kadri.Digital.Library.Management.System.entity.User;
+import kadri.Digital.Library.Management.System.exception.UserNotFoundException;
 import kadri.Digital.Library.Management.System.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -22,11 +23,11 @@ public class UserServiceImpl implements UserService{
         String name = oAuth2User.getAttribute("name");
         String avatarUrl = oAuth2User.getAttribute("avatar_url");
 
-        Optional<User> existingUser = userRepository.findByUsername(username);
-        if(existingUser.isPresent()) return existingUser.get();
-
-        User newUser = new User(username, name, avatarUrl, "USER");
-        return userRepository.save(newUser);
+        return userRepository.findByUsername(username)
+                .orElseGet(() -> {
+                    User newUser = new User(username, name, avatarUrl, "USER");
+                    return userRepository.save(newUser);
+                });
     }
 
     @Override
@@ -37,19 +38,25 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public void delete(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException("User with ID " + userId + " not found.");
+        }
         userRepository.deleteById(userId);
     }
 
     @Override
     public String getUser(Long id) {
-        Optional<User> user = userRepository.findById(id);
-        if(user.isPresent())
-        return user.get().getUsername();
-        else return null;
+        return userRepository.findById(id)
+                .map(User::getUsername)
+                .orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found."));
     }
 
     @Override
     public Optional<User> findByUsername(String username) {
-        return Optional.of(userRepository.findByUsername(username).get());
+        return userRepository.findByUsername(username)
+                .or(()->{
+                    throw new UserNotFoundException("User with username " + username + " not found.");
+                });
     }
+
 }
